+++ docker/init-docker.php
@@ -1,10 +1,16 @@
 <?php
 
+use Monolog\Logger;
+use Pastell\Database\DatabaseUpdater;
+
 try {
     echo "Initialisation de Pastell [DOCKER]\n";
 
 # Première étape : Sans la connexion BD vu que celle-ci n'existe pas encore...
     require_once(__DIR__ . "/../init-no-db.php");
+    /**
+     * @var Logger $logger
+     */
 
     echo "Utilisation de la base " . BD_DSN . " avec l'utilisateur " . BD_USER . "\n";
 
@@ -13,19 +19,14 @@ try {
     $sqlQuery->waitStarting(function ($message) {
         echo "[" . date("Y-m-d H:i:s") . "][Pastell - wait for MySL] $message\n";
     });
-    $databaseUpdate = new DatabaseUpdate(file_get_contents(__DIR__ . "/../installation/pastell.bin"), $sqlQuery);
-    $databaseUpdate->majDatabase(
-        $sqlQuery,
-        function ($message) {
-            echo "[" . date("Y-m-d H:i:s") . "][Pastell - SQL init] $message\n";
-        }
-    );
+    $databaseUpdater = new DatabaseUpdater($sqlQuery, $logger);
+    $databaseUpdater->update();
 
 # Deuxième étape : initialisation normal de Pastell
     require_once __DIR__ . "/../init.php";
 
     /** @var PastellBootstrap $pastellBootstrap */
-    $pastellBootstrap = $objectInstancier->getInstance(PastellBootstrap::class);
+    $pastellBootstrap = ObjectInstancierFactory::getObjetInstancier()->getInstance(PastellBootstrap::class);
 
     $envWrapper = new EnvWrapper();
     $utilisateurObject = new UtilisateurObject();